--- a/spandex-common/src/main/scala/com/socrata/spandex/common/client/ResponseExtensions.scala
+++ b/spandex-common/src/main/scala/com/socrata/spandex/common/client/ResponseExtensions.scala
@@ -89,7 +89,8 @@ case class FieldValue(datasetId: String,
                       value: String) {
   lazy val docId = FieldValue.makeDocId(datasetId, copyNumber, columnId, rowId)
   lazy val compositeId = FieldValue.makeCompositeId(datasetId, copyNumber, columnId)
-  lazy val completionValue = CompletionValue(value)
+  // *sigh* a cluster side analysis char_filter doesn't catch this one character in time.
+  lazy val completionValue = CompletionValue(value.replaceAll("\u001f", ""))
 
   // Needed for codec builder
   def this(datasetId: String,