--- a/spandex-common/src/test/scala/com/socrata/spandex/common/client/SpandexElasticSearchClientSpec.scala
+++ b/spandex-common/src/test/scala/com/socrata/spandex/common/client/SpandexElasticSearchClientSpec.scala
@@ -4,7 +4,7 @@ import com.socrata.datacoordinator.secondary.LifecycleStage
 import com.socrata.spandex.common.client.ResponseExtensions._
 import com.socrata.spandex.common.{SpandexConfig, TestESData}
 import org.elasticsearch.action.index.IndexRequestBuilder
-import org.elasticsearch.common.unit.Fuzziness
+import org.elasticsearch.index.engine.IndexFailedEngineException
 import org.scalatest.{BeforeAndAfterAll, BeforeAndAfterEach, FunSuiteLike, Matchers}
 
 // scalastyle:off
@@ -305,5 +305,32 @@ class SpandexElasticSearchClientSpec extends FunSuiteLike with Matchers with Bef
     a[IllegalArgumentException] shouldBe thrownBy {
       client.datasetCopyLatest(datasets(0), Some(Number(42))).get.copyNumber
     }
+  }
+
+  // previously, a blank value rendered the following json and caused an exception in lucene
+  // this is allegedly fixed in elasticsearch 1.5.3
+  // we could have fixed it by filtering out empty strings explicitly
+  // completion pre-analysis handles empty string by happy accident
+  test("handle value string is empty or null") {
+    try {
+      // org.elasticsearch.index.engine.IndexFailedEngineException: [spandex][2] Index failed for [field_value#primus.1234|2|3|60]
+      // Cause: java.lang.IllegalStateException: from state (0) already had transitions added
+      client.client.prepareIndex(config.es.index, config.es.fieldValueMapping.mappingType, "primus.1234|2|3|60")
+        .setSource(
+          """{
+            | "composite_id":"primus.1234|2|3",
+            | "copy_number":2,
+            | "dataset_id":"primus.1234",
+            | "row_id":60,
+            | "value":"",
+            | "column_id":3
+            |}""".
+            stripMargin)
+        .setRefresh(true).execute.actionGet
+    } catch {
+      case e: IndexFailedEngineException => // expected on previous versions
+    }
+
+    client.indexFieldValue(FieldValue(datasets(0), 1L, 2L, 61L, ""), refresh = true)
   }
 }