What's new
--------
+ Empty string should be fixed in https://github.com/elastic/elasticsearch/pull/11158 version 1.5.3 onward.
+ Control character unit separator (\x1F) is forbidden.

Performance benchmarks
--------
I am wary of adding a string replacement (regex) to the FieldValue constructor. Here is a performance estimate of the changes. No significant impact observed.

Before
```
Benchmark                                                                Mode  Cnt    Score    Error  Units
c.s.s.common.CompletionAnalyzerBenchmark.analyzeCompletionTokens         avgt   16  117.379 ±  2.723  us/op
c.s.s.secondary.ESIndexBenchmark.indexKeyword                            avgt   16  180.698 ± 21.023  ms/op
c.s.s.secondary.ESIndexBenchmark.indexPhrase                             avgt   16  202.310 ± 23.924  ms/op
```
After - string replacement in `new FieldValue()` constructor
```
Benchmark                                                                Mode  Cnt    Score    Error  Units
c.s.s.common.CompletionAnalyzerBenchmark.analyzeCompletionTokens         avgt   16  116.805 ±  5.649  us/op
c.s.s.secondary.ESIndexBenchmark.indexKeyword                            avgt   16  178.085 ± 28.357  ms/op
c.s.s.secondary.ESIndexBenchmark.indexPhrase                             avgt   16  194.008 ± 21.316  ms/op
```
After - string replacement in `RowOpsHandler.fieldValueFromDatum()` function
```
Benchmark                                                                Mode  Cnt    Score    Error  Units
c.s.s.common.CompletionAnalyzerBenchmark.analyzeCompletionTokens         avgt   16  115.659 ±  2.374  us/op
c.s.s.secondary.ESIndexBenchmark.indexKeyword                            avgt   16  175.888 ± 19.800  ms/op
c.s.s.secondary.ESIndexBenchmark.indexPhrase                             avgt   16  183.786 ±  9.351  ms/op
```