@cloudbees-pull-request-builder 2015-05-28
[spandex #330](https://socrata-oss.ci.cloudbees.com/job/spandex/330/) SUCCESS
This pull request looks good