@cloudbees-pull-request-builder 2015-05-28
[spandex #331](https://socrata-oss.ci.cloudbees.com/job/spandex/331/) SUCCESS
This pull request looks good