@sokrahta 2015-05-28
why Jenkins why...
> Could not create directory /data/jenkins/workspace/sbt-build/spandex-http/target/streams/$global/projectDescriptors/$global/streams