@cloudbees-pull-request-builder 2015-05-28
[spandex #333](https://socrata-oss.ci.cloudbees.com/job/spandex/333/) SUCCESS
This pull request looks good