@urmilan 2015-05-28
:shipit: 
Try blowing away the workspace if Jenkins isn't cooperating.